Add pub date to entries
Since there doesn't seem to be a publish date on the scraped web site
data, we use the trick of comparing the entries with the ones
previously scraped into challenges.json. If the entry is new, then the
pub date is the timestamp of the scrape. Note this is fragile because
currently the method of determining a unique ID for entries is their
URL, which has been observed to change, on occasion.